refactor: Add clearBorrowedHandler to Borrowed tab
This commit adds a clearBorrowedHandler function to the Borrowed tab component. The function is used to clear the list of borrowed books and display a success alert to the user. This enhances the usability of the Borrowed tab and allows users to easily remove all borrowed books.

Refactor the Borrowed tab component to include the clearBorrowedHandler function and update the necessary imports and dependencies.